cgh ! mapping app, rebase, tests
