+++ src/country_workspace/admin/__init__.py
@@ -12,7 +12,6 @@ from .household import HouseholdAdmin
 from .individual import IndividualAdmin
 from .job import AsyncJobAdmin
 from .locations import AreaAdmin, AreaTypeAdmin, CountryAdmin
-from .mapping import FieldMappingRuleAdmin, MappingProfileAdmin
 from .office import OfficeAdmin
 from .program import ProgramAdmin
 from .rdp import RdpAdmin
@@ -37,10 +36,8 @@ __all__ = [
     "BeneficiaryGroupAdmin",
     "ConstanceAdmin",
     "CountryAdmin",
-    "FieldMappingRuleAdmin",
     "HouseholdAdmin",
     "IndividualAdmin",
-    "MappingProfileAdmin",
     "OfficeAdmin",
     "ProgramAdmin",
     "RdpAdmin",
