+++ src/country_workspace/mapping/admin/__init__.py
@@ -1,0 +1,8 @@
+from .field_mapping_rule import FieldMappingRuleAdmin
+from .profile import MappingProfileAdmin
+
+
+__all__ = [
+    "FieldMappingRuleAdmin",
+    "MappingProfileAdmin",
+]
