+++ src/country_workspace/mapping/admin/profile.py
@@ -1,9 +1,10 @@
 from django.contrib import admin
+from django import forms
 from django.db import models
 from django.http import HttpRequest
 
-from ..models import MappingProfile, FieldMappingRule
-from .base import BaseModelAdmin
+from country_workspace.mapping.models import MappingProfile
+from country_workspace.admin.base import BaseModelAdmin
 
 
 @admin.register(MappingProfile)
@@ -33,26 +34,9 @@ class MappingProfileAdmin(BaseModelAdmin):
     def get_queryset(self, request: HttpRequest) -> models.QuerySet[MappingProfile]:
         return super().get_queryset(request).select_related("parent", "created_by").prefetch_related("program")
 
-
-@admin.register(FieldMappingRule)
-class FieldMappingRuleAdmin(BaseModelAdmin):
-    list_display = ("name", "profile", "order", "is_active")
-    list_filter = ("profile", "is_active")
-    search_fields = ("name", "description", "profile__name")
-    ordering = ("profile", "order", "name")
-    fields = (
-        "name",
-        "description",
-        "profile",
-        "expression",
-        "order",
-        "is_active",
-        "created_by",
-        "created_at",
-        "updated_at",
-    )
-
-    readonly_fields = ("created_by", "created_at", "updated_at")
-
-    def get_queryset(self, request: HttpRequest) -> models.QuerySet[FieldMappingRule]:
-        return super().get_queryset(request).select_related("profile__parent", "created_by")
+    def save_model(
+        self, request: HttpRequest, obj: MappingProfile, form: forms.ModelForm | None = None, change: bool = False
+    ) -> None:
+        if not change:
+            obj.created_by = request.user
+        super().save_model(request, obj, form, change)
