+++ src/country_workspace/mapping/models/__init__.py
@@ -1,0 +1,7 @@
+from .field_mapping_rule import FieldMappingRule
+from .profile import MappingProfile
+
+__all__ = [
+    "FieldMappingRule",
+    "MappingProfile",
+]
