+++ tests/utils/test_utils_fields.py
@@ -12,7 +12,6 @@ from country_workspace.utils.fields import (
     clean_field_name,
     TO_REMOVE_VALUES,
     clean_field_names,
-    map_fields,
     extract_uuid,
 )
 from country_workspace.utils.flex_fields import Base64ImageInput, Base64ImageField, ConsentSharingChoice
@@ -37,20 +36,6 @@ def test_clean_field_names(mocker: MockerFixture) -> None:
 
     assert cleaned == {clean_field_name_mock.return_value: "bar"}
     clean_field_name_mock.assert_called_once_with(key)
-
-
-@pytest.mark.parametrize(
-    ("input_fields", "expected_output"),
-    [
-        ({"gender": "male"}, {"sex": "male"}),
-        ({"name": "John"}, {"name": "John"}),
-        ({}, {}),
-        ({"gender": "female", "age": "30"}, {"sex": "female", "age": "30"}),
-    ],
-)
-def test_map_fields(input_fields, expected_output):
-    result = map_fields(input_fields)
-    assert result == expected_output
 
 
 @pytest.mark.parametrize("value", [None, "", "test"])
